bring all logic to parents component in signature popup
