+++ apps/web/src/screens/Request/Signature/index.tsx
@@ -1,10 +1,18 @@
-import { useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { type SlideOption, Slider } from '@walless/gui';
-import { type ResponsePayload } from '@walless/messaging';
-import { handleRequestSignature } from 'bridge/listeners';
+import {
+	type MessagePayload,
+	type RequestType,
+	type ResponsePayload,
+	PopupType,
+} from '@walless/messaging';
+import {
+	getMessageOrTransaction,
+	handleRequestSignature,
+} from 'bridge/listeners';
 
-import { type PayloadOptions } from '../shared';
+import { type PayloadOptions, initializeKernelConnect } from '../shared';
 
 import RequestSignatureApproval from './Approval';
 import RequestSignaturePasscode from './Passcode';
@@ -12,34 +20,40 @@ import RequestSignaturePasscode from './Passcode';
 export const RequestSignature = () => {
 	const { requestId = '' } = useParams();
 	const [activeIndex, setActiveIndex] = useState(0);
+	const [payload, setPayload] = useState<MessagePayload>();
 	const options = useRef<PayloadOptions>({
 		sourceRequestId: requestId,
 		isApproved: false,
 		passcode: '',
 	});
 
-	const handleDenyRequest = () => {
-		handleRequestSignature(options.current);
-	};
+	const handleDenyRequest = useCallback(async () => {
+		handleRequestSignature(options.current, payload?.type as RequestType);
+	}, [payload]);
 
 	const handleApproveRequest = () => {
 		options.current.isApproved = true;
 		setActiveIndex(1);
 	};
 
-	const handleResolveRequest = async (
-		passcode: string,
-	): Promise<ResponsePayload> => {
-		options.current.passcode = passcode;
+	const handleResolveRequest = useCallback(
+		async (passcode: string): Promise<ResponsePayload> => {
+			options.current.passcode = passcode;
 
-		return await handleRequestSignature(options.current);
-	};
+			return await handleRequestSignature(
+				options.current,
+				payload?.type as RequestType,
+			);
+		},
+		[payload],
+	);
 
 	const sliderItems: SlideOption[] = [
 		{
 			id: 'approval',
 			component: () => (
 				<RequestSignatureApproval
+					content={payload?.message || payload?.transaction}
 					onDeny={handleDenyRequest}
 					onApprove={handleApproveRequest}
 				/>
@@ -55,6 +69,16 @@ export const RequestSignature = () => {
 			),
 		},
 	];
+
+	useEffect(() => {
+		initializeKernelConnect(`${PopupType.SIGNATURE_POPUP}/${requestId}`);
+	}, []);
+
+	useEffect(() => {
+		getMessageOrTransaction(requestId).then((result) => {
+			setPayload(result as MessagePayload);
+		});
+	}, [requestId]);
 
 	return (
 		<Slider
